+++ backend/app/models/search_alert.py
@@ -1,6 +1,7 @@
 from datetime import datetime
 from typing import Optional
-from sqlalchemy import String, Boolean, Float, DateTime, Integer, ForeignKey, func, Index
+from decimal import Decimal
+from sqlalchemy import String, Boolean, Numeric, DateTime, Integer, ForeignKey, func, Index
 from sqlalchemy.orm import Mapped, mapped_column
 from sqlalchemy import Enum as SQLEnum
 from app.models.enums import SearchAlertStatus
@@ -19,7 +20,7 @@ class SearchAlert(Base):
     )
     category: Mapped[Optional[str]] = mapped_column(String(50), nullable=True)
     query: Mapped[Optional[str]] = mapped_column(String(200), nullable=True)
-    max_price: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
+    max_price: Mapped[Optional[Decimal]] = mapped_column(Numeric(12, 2), nullable=True)
     status: Mapped[SearchAlertStatus] = mapped_column(SQLEnum(SearchAlertStatus, values_callable=lambda obj: [e.value for e in obj]), default=SearchAlertStatus.ACTIVE, nullable=False, index=True)
     created_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True), server_default=func.now()
